Move graph.py to externalapis
